+++ pyrgg/params.py
@@ -5,14 +5,14 @@ import os
 MENU_ITEMS1 = {
     "file_name": "- File Name : ",
     "output_format": "- Graph Formats : \n1- DIMACS(.gr)\n2- JSON(.json)\n3- CSV(.csv)\n4- YAML(.yaml)\n5- WEL(.wel)\n6- ASP(.lp)\n7- Pickle(.p)\n8- UCINET DL(.dl)\n9- TGF(.tgf)\n10- TSV("
-                     ".tsv)\n11- Matrix Market(.mtx)\n12- Graph Line(.gl)\n13- GDF(.gdf)\n14- GML(.gml)\n",
+                     ".tsv)\n11- Matrix Market(.mtx)\n12- Graph Line(.gl)\n13- GDF(.gdf)\n14- GML(.gml)\n15- GEXF(.gexf)\n",
     "weight": "- Weighted[1] or Unweighted[2]"}
 
 MENU_ITEMS2 = {"vertices": "- Vertices Number : ",
                "max_weight": "- Max Weight : ",
                "min_weight": "- Min Weight : ",
-               "min_edge": "- Min Edge Number :",
-               "max_edge": "- Max Edge Number :",
+               "min_edge": "- Min Edge Number : ",
+               "max_edge": "- Max Edge Number : ",
                "sign": "- Signed[1] or Unsigned[2]",
                "direct": "- Directed[1] or Undirected[2]",
                "self_loop": "- Self Loop[1] or No Self Loop[2]",
@@ -33,7 +33,8 @@ SUFFIX_MENU = {
     11: ".mtx",
     12: ".gl",
     13: ".gdf",
-    14: ".gml"}
+    14: ".gml",
+    15: ".gexf"}
 
 
 PYRGG_VERSION = "0.8"